Cambio de nombre de algunas clases
